% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_values_in_list}
\alias{check_values_in_list}
\title{Check if values are in list}
\usage{
check_values_in_list(values, check_list)
}
\arguments{
\item{values}{Values to check}

\item{check_list}{Values to check 'values' against}
}
\description{
Check if values are in list
}
