% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_vars_exist}
\alias{check_vars_exist}
\title{Columns exist validation}
\usage{
check_vars_exist(data = NULL, vars = NULL)
}
\arguments{
\item{data}{Data}

\item{vars}{Variables to check if they exist in Data}
}
\description{
Checks if the columns in data exist.
}
