% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{create_model_matrix}
\alias{create_model_matrix}
\title{Create model matrix for regression}
\usage{
create_model_matrix(data = NULL, intercept = TRUE)
}
\arguments{
\item{data}{Data}

\item{intercept}{If intercept should be included in the model}
}
\value{
matrix
}
\description{
Function for creating a model matrix of the input data
}
