% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_missing_summary}
\alias{get_missing_summary}
\title{Missing value summary}
\usage{
get_missing_summary(data)
}
\arguments{
\item{data}{Data}
}
\value{
list
}
\description{
Get summary of the missing values of data
}
