% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear regression}
\usage{
linear_regression(data = NULL, y = NULL, covariates = NULL, intercept = TRUE)
}
\arguments{
\item{data}{Data}

\item{y}{Name of y variable to predict}

\item{covariates}{Variables to include in the model}

\item{intercept}{Whether to include intercept in the model}
}
\value{
list
}
\description{
Function for computing a standard linear regression
}
