% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic regression}
\usage{
logistic_regression(
  data = NULL,
  y = NULL,
  covariates = NULL,
  intercept = TRUE,
  max_iter = 100,
  tol = 1e-06
)
}
\arguments{
\item{data}{Data}

\item{y}{The binary value to estimate}

\item{covariates}{The variables used as covariates}

\item{intercept}{If intercept should be included}

\item{max_iter}{Maximum number of iterations}

\item{tolerance}{Tolerance of difference in estimation of coefficients}
}
\value{
list Contains the model
}
\description{
Creates a logistic regression model
}
