% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{normality_test}
\alias{normality_test}
\title{Normality test}
\usage{
normality_test(x = NULL, tests = c("shapiro_wilk", "jarque_bera")[1])
}
\arguments{
\item{x}{Vector in values to be used for test}

\item{tests}{Which tests to use}
}
\description{
Function for normality tests
}
