% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print.aesn_linear_regression}
\alias{print.aesn_linear_regression}
\title{Print summary for Linear regression}
\usage{
\method{print}{aesn_linear_regression}(model, ...)
}
\arguments{
\item{object}{aesn_linear_regression}
}
\description{
Function printing the summary of the fitted model
}
